upgrade RequireJS, make its optimizer happy with socket.io.
The optimizer really wants to look inside socket.io for dependencies
even though it is not responsible for grouping it, so we tell it about a
basically empty file so it is happy.  We are not having it bundle
socket.io right now just do decouple the npm stuff for now.  However,
given that the newest npm does local install stuff and we rebuild our
optimized file every run, it might make sense to bundle it in somehow.