deuxdrop loggest checkpoint; we can process loggest dumps and show them
Where "show" in this case means that when you click on the failure you
see a JSON.stringify(obj, null, 2) text presentation of the test case's
logger alongside the schema dump.

Next step is to get the loggest log processing being smarter.  That will
happen in ui-loggest.js and chew-loggest.js, both of which are just
comments and copied and pasted code right now.