automatically expand steps with any errors inside, give red border
Sometimes steps pass even though something broke; this helps it make
more obvious that something bad happened, even if the test itself
doesn't fully pick up on it.

A possible limitation of this UI-only approach is that the summary does
not properly convey that something bad might have happened.  Since the
preprocessing step uses chew-loggest, this can be easily addressed.

The flip side is that in cases where errors are expected, this could get
annoying.  We presumably could use the presence of an expectation
wrapping the error or the like to mark it inert...