TEST_ONLY_asyncJobs supported.  specialized styling
Right now I'm cramming the IMAP connection data through the test-only args
mechanism.  This is super-chatty, so I am trying to cause it to be
visually filterable a bit by reducing its contrast.  It would probably be
better to just support a mode where the data is initially collapsed into
an object that you have to click on to see.