Gray background for log entries outside the step
In a step we actually display the log entries between the previous step
and our step, as well as during our step.  And if we're the last step,
after the step.  We weren't visually distinguishing these before.  I am
investigating a weird intermittent failure that justified having an ugly
gray background to help make this case more clear.  Sadly, I still am not
sure what is going on...